refactor out company table into hook
